requirements.txt
flask
jinja2
pandas
tqdm
python-dotenv
flask-session
flask-wtf
email-validator
gunicorn
pymongo[srv]>=4.6
pydantic>=2,<3